todo_list.js
I optimized this code and combined the "add()" and "del()" functions into a single function, which will be called "createListElement()". This would simplify the code structure and make it easier to remember the function used to add and remove elements from the list.

Here, the "createListElement()" function builds the new list element by creating an "li" tag and a "span" tag for each new element added to the list. After this, the function adds the new element to the list and also adds two events that occur when a click happens on the new list element and on the "span" tag to remove the element. Finally, the function clears the text in the input field and adds an event listener to the input field that allows the user to add new elements by pressing the Enter key.